requirements.txt
streamlit>=1.30.0
pytesseract>=0.3.10
easyocr>=1.7.0
pymupdf>=1.24.0
Pillow>=10.0.0
opencv-python-headless>=4.8.0
scikit-learn>=1.3.0
sentence-transformers>=2.2.0
python-Levenshtein>=0.21.0
google-generativeai>=0.8.0
transformers>=4.35.0
torch>=2.1.0
numpy>=1.24.0
pandas>=2.0.0
